backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
fonttools>=4.55.0
freetype-py>=2.4.0
Pillow>=10.0.0
